Stream::process: process stream values in parallel

+++ test/util/async/test_stream_filter.rs
@@ -9,6 +9,6 @@ pub fn test_stream_map_filter() {
 
     s.each(move |i| tx.send(i).unwrap()).fire();
 
-    let vals: Vec<i32> = rx.iter().collect();
+    let vals: Vec<usize> = rx.iter().collect();
     assert_eq!([0, 2, 4].as_slice(), vals.as_slice());
 }
